fix: 글 등록 페이지에서 타이핑 할 때 마다 api 호출되는 문제 수정
